+++ CHANGELOG.md
@@ -23,6 +23,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix 1579: Add ImageChange triggers to application container when resource fragments are used
 * Fix 1607: Fix YAML aggregation for OpenShift
 * Refactor 1592: Removed Deprecated/Unused Osio specific enrichers.
+* Fix 1608: Added Enricher for namespace to project conversion.
 
 ### 4.0.0 (14-03-2019)
 * Refactor 678: Enricher Workflow to allow direct generation of OpenShift objects.